Update pingid.py to handle non jwt response
When calling 'rest/4/getorgreport/do' endpoint, the content-type of the response is 'application/octet-stream', not 'application/json', so the jwt.decode() fails. This update will check the headers for content-type, and if it is 'application/octet-stream' it will set the extracted_response to r.text, otherwise it will set extracted response to jwt.decode(r.content, key, algorithms=['HS256']) like it currently does.