Create getOrgReport.py
This is a sample script to get the organization report from pingid.

It hits the 'rest/4/createjob/do' endpoint to create the job, the 'rest/4/getjobstatus/do' endpoint to check job status, and then the 'rest/4/getorgreport/do' to get the report. It then creates a csv file of the organization report.